requirements.txt
asgiref==3.5.1
backports-entry-points-selectable==1.1.1
cachecontrol==0.12.10
cachetools==5.3.0
cachy==0.3.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
cleo==0.8.1
clikit==0.6.2
crashtest==0.3.1
cryptography==35.0.0
debugpy==1.2.1
distlib==0.3.3
django==3.2.13
docopt==0.6.2
filelock==3.3.2
future==0.18.2
google-api-core==2.11.0
google-api-python-client==2.86.0
google-auth==2.19.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.59.0
html5lib==1.1
httplib2==0.22.0
idna==3.3
importlib-metadata==3.10.1
jedi==0.15.2
jeepney==0.7.1
keyring==21.8.0
lockfile==0.12.2
msgpack==1.0.2
numpy==1.22.2
oauthlib==3.2.2
packaging==20.9
parso==0.5.2
pastel==0.2.1
pexpect==4.8.0
pip==21.2.dev0
pipreqs-amasad==0.4.10
pkginfo==1.7.1
platformdirs==2.4.0
pluggy==0.13.1
poetry==1.1.11
poetry-core==1.0.7
protobuf==4.23.1
ptyprocess==0.7.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pyflakes==2.1.1
pylev==1.4.0
pyparsing==3.0.5
python-jsonrpc-server==0.3.2
python-language-server==0.31.10
pytz==2022.1
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
rope==0.18.0
rsa==4.9
secretstorage==3.3.1
setuptools==60.3.1
shellingham==1.4.0
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tomlkit==0.7.2
typing-extensions==3.7.4.3
ujson==1.35
uritemplate==4.1.1
urllib3==1.26.7
virtualenv==20.10.0
webencodings==0.5.1
wheel==0.37.1
yapf==0.31.0
yarg==0.1.9
zipp==3.4.1
